[CMake] Add LSP provider using the cmake project (#11)
Most projects using MLIR use it as a submodule. When working with these
projects it is then desirable to use the `tblgen-lsp-server` built as
part of the project.

This PR adds such an implementation that scans CLions cmake workspace
for the target and executable location. A lot of TODOs are left for the
future such as build configuration selection or prompting users to build
the executable if not yet done.