[CMake] Fix race condition and cancellations at startup
The current implementation of fetching the cmake build profile has two large issues:
* Since service initialization is lazy, the listeners may only start receiving events as soon as the service is requested for the first time
* Since the implementation used a cancellable blocking coroutine await, it is likely to be cancelled and needs an explicit restart

The new implementation leverages the declarative listeners to fix the first issue. It guarantees that the given listener classes will receive the events by lazy initializing them on first publish.
Additionally, the LSP is restarted once the run manager is initialized and the cmake build profile known.