[TableGen] Add basic file type and lexer
This PR adds the initial skeleton needed to add syntax highlighting for TableGen in IntelliJ.
To do so, a very basic lexer and parser have been implemented that never generates error tokens.
The parser simply accepts every possible token kind and is used to generate the lexer token definitions.
The lexer is used for the actual syntax highlighting.

More tokens will be added to the lexer in the future but the parser may be kept as simple as is reasonable.
Instead, features such as [semantic tokens support](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_semanticTokens) should probably be added to the LSP instead.