[TableGen] Implement basic parser
This PR is an initial implementation of a parser and PSI for TableGen within IntelliJ. It is already capable of parsing all MLIR TableGen files that I tested it against and build and accurate AST useful for e.g. syntax highlighting.

The reason for implementing this rather than enhancing the LSP is mainly the difficulty of pushing some of these changes upstream, especially as the datastructures are optimized for fast and low-memory compilation, not IDE-like use-cases.

The parser does currently not have good error recovery which is left to future PRs.
Tests are not 100% complete either.