[TableGen] Add context state to `TableGenFile`
A tablegen files context is used for rarely-changing global state that has to nevertheless be computed from all project files.

The current context is used to calculate and propagate the include paths used to compile table gen files and required to resolve `include` statements.
Having this state replaces the use of `TableGenIncludeGraph` and the corresponding index.