[LSP] Disable all but diagnostic
At this point the IDE is more capable at performing the majority of these tasks.
Some of these, such as find usages, show up as an additional menu, making it confusing to users which functionality is provided by the LSP or by the IDE.

Disable all but diagnostics support in the LSP for now. Diagnostics is likely to be the last thing that will be replaced if ever