[TableGen] Amortized identifier lookup with stub support (#201)
The previous local identifier lookup always performed a walk of the Psi
and was therefore always linear in complexity and did not support stubs
well.

The new lookup now calculates a scope map which in most cases will be
`O(1)`, or `O(log n)` where `n` is the number of elements with the same
name (i.e. usually tiny).